issue #1 autoenv masks return value of builtin cd
At autovenv.sh:53

cd() {
    echo $@
    builtin cd "$@"
    autovenv
}

autovenv is called regardless of whether or not builtin cd "$@" succeeds;
this masks the return value of builtin cd. This breaks any checks for
the return value of cd subsequent to sourcing autovenv.sh, e.g.

cd some/directory/that/does/not/exist && rm -rf important_directory/ will
remove important_directory/ in the current directory because the call to
builtin cd failed, but autovenv was called afterwards, and it apparently
always returns true, so cd() is successful, even though builtin cd is not.

Making the call to autovenv conditional on the success of builtin cd
will fix this:

cd() {
    echo $@
    builtin cd "$@" && autovenv
}